backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0

# Stock Data Sources
yfinance==0.2.36  # Yahoo Finance (primary)
aiohttp==3.9.1    # Alpha Vantage & Polygon.io APIs

# Data Processing
pandas==2.2.0
numpy==1.26.3
lxml==5.1.0
beautifulsoup4==4.12.3

# Database & Caching
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1

# Background Jobs
apscheduler==3.10.4

# Configuration
python-dotenv==1.0.0
